feat: create @decepticon/streaming shared package for Web/CLI
Extract duplicated streaming types, constants, and utilities from
Web (useChat.ts) and CLI (useAgent.ts) into a shared workspace package.

- Create packages/streaming with SubagentCustomEvent, STREAM_OPTIONS,
  extractText(), stripResultTags(), SubAgentSession, deriveSubAgentSessions()
- Wire Web and CLI to import from @decepticon/streaming
- Remove all inline duplicates from both clients
- Add clients/web to npm workspaces
- Align CLI SDK version to ^1.8.8 (matching Web)
- Re-export SubAgentSession from CLI types.ts via shared package
