+++ packages/streaming/package.json
@@ -1,0 +1,15 @@
+{
+  "name": "@decepticon/streaming",
+  "version": "0.1.0",
+  "description": "Shared LangGraph streaming types, constants, and utilities for Web and CLI",
+  "type": "module",
+  "main": "src/index.ts",
+  "types": "src/index.ts",
+  "scripts": {
+    "typecheck": "tsc --noEmit"
+  },
+  "dependencies": {},
+  "devDependencies": {
+    "typescript": "^5.7.0"
+  }
+}
